+++ beesdoo_purchase/readme/DESCRIPTION.rst
@@ -1,5 +1,1 @@
-Extends Purchase module:
-
-- A filter w.r.t. the main sellers is placed on the products field of a
-  purchase order.
-- Allows adapting a product's purchase and/or selling price from a PO.
+Emptied.
