+++ sale_suggested_price/views/purchase_order.xml
@@ -1,0 +1,73 @@
+<odoo>
+    <data>
+        <record model="ir.ui.view" id="purchase_order_form">
+            <field name="name">purchase_order_form</field>
+            <field name="model">purchase.order</field>
+            <field name="inherit_id" ref="purchase.purchase_order_form" />
+            <field name="arch" type="xml">
+                <xpath expr="//header" position="inside">
+                     <button
+                        name="button_adapt_price"
+                        states="draft,sent,purchase"
+                        string="Adapt Prices"
+                        type="object"
+                    />
+                </xpath>
+                <xpath expr="//div[hasclass('oe_button_box')]" position="inside">
+                    <button
+                        name="action_select_deselect_adapt_purchase_price"
+                        type="object"
+                        attrs="{'invisible': [('select_all_purchase_price', '=', False)]}"
+                        string="Deselect All Is Purchase Price"
+                        icon="fa-square-o"
+                        class="oe_stat_button"
+                        help="Deselect All Is Purchase Price checkboxes to adapt the purchase price on the product page when confirming Purchase Order"
+                    />
+                    <button
+                        name="action_select_deselect_adapt_purchase_price"
+                        type="object"
+                        attrs="{'invisible': [('select_all_purchase_price', '=', True)]}"
+                        string="Select All Is Purchase Price"
+                        icon="fa-check-square"
+                        class="oe_stat_button"
+                        help="Select All Is Purchase Price checkboxes to adapt the purchase price on the product page when confirming Purchase Order"
+                    />
+                    <button
+                        name="action_select_deselect_adapt_selling_price"
+                        type="object"
+                        attrs="{'invisible': [('select_all_selling_price', '=', False)]}"
+                        string="Deselect All Is Selling Price"
+                        icon="fa-square-o"
+                        class="oe_stat_button"
+                        help="Deselect All Is Selling Price checkboxes to adapt the selling price on the product page when confirming Purchase Order"
+                    />
+                    <button
+                        name="action_select_deselect_adapt_selling_price"
+                        type="object"
+                        attrs="{'invisible': [('select_all_selling_price', '=', True)]}"
+                        string="Select All Is Selling Price"
+                        icon="fa-check-square"
+                        class="oe_stat_button"
+                        help="Select All Is Selling Price checkboxes to adapt the selling price on the product page when confirming Purchase Order"
+                    />
+                </xpath>
+                <field name="date_order" position="after">
+                    <field name="select_all_purchase_price" invisible="1" />
+                    <field name="select_all_selling_price" invisible="1" />
+                </field>
+                <field name="product_id" position="attributes">
+                    <attribute name="domain">[
+                        ('main_seller_id','=', parent.partner_id),
+                        ('purchase_ok', '=', True) ]
+                    </attribute>
+                </field>
+                <field name="price_subtotal" position="after">
+                    <field name="adapt_purchase_price" string="Is Purchase Price" />
+                    <field name="adapt_selling_price" string="Is Selling Price" />
+                </field>
+            </field>
+        </record>
+
+
+    </data>
+</odoo>
